Add files via upload
This toolbox is a plug-in to the open-source code OpenFoam. The toolbox is originated from waves2Foam which is based on external wave theories and expand it by incorporating internal wave theory. Using this toolbox, one can generate regular and irregular waves using both the external and internal wave theories and minimize the effect by the waves refected from the structure by implementing enegy dissipation zone.